+++ CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+### Added
+- `JsonapiResponse` exporting
+
+
 ## [1.0.0-x.14] - 2019-02-21
 ### Added
 - KeyServerCaller
 
+### Changed
+- `Wallets._getSigners` method according to the new API
+
 ### Fixed
-- Wallet.fromEncrypted in cases when you provide seed only
+- `makeChangeSignerTransaction` function according to the signer roles
 
 ## [1.0.0-x.13] - 2019-02-21
 ### Added